I just added the ability to specify an "absolute:" path to the templates directory.
It works something like this:

services.VelocityService.file.resource.loader.path =
@ABSOLUTE_PATH@/templates,@ABSOLUTE_PATH@/templates/flux

You can then use Ant to do the filtering of the TR.props file:

        <filter token="ABSOLUTE_PATH" value="absolute:${src.dir}"/>
        <copy todir="${build.project.webinf}/conf" filtering="yes">
            <fileset dir="${src.conf.dir}/" defaultexcludes="no">
                <include name="*.properties"/>
                <include name="*.xml"/>
            </fileset>
        </copy>

This has been tested and works wonderfully!

What this means is that your build system can define an absolute path and
you don't need to copy the templates into the webapp's directory anymore!
You can simply point at them in any directory for development purposes.
This makes it easier for designers because the templates directory doesn't
need to be nested way down in a complicated directory structure. It also
removes the need to have a "build" step in order to copy the modified
templates into another directory.

I'm open to suggestions for something other than "absolute:", but this
works for me for now.

woo hoo! I have been wanting to solve this problem for a long while now.

-jon


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222113 13f79535-47bb-0310-9956-ffa450edef68