patch by 	Stephen Haberman <stephenh@chase3000.com>
The Fulcrum maven:clean target also calls Fulcrum's own clean target.
However, it's broken because it refers to the old services dir and also
it un-needed because all of the Torque generation is done into the
prepareSourceDirectory, which gets deleted along with the entire target
dir by maven.

I've attached a simple patch that gets ride of the clean target and
remove's maven:cleans' dependency on it.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222508 13f79535-47bb-0310-9956-ffa450edef68