Removed worthless <mkdir> from the "check-om" goal.
Currently, the Fulcrum Maven build must be run _twice_ after a "clean"
due a problem in Maven's driver.jelly (jvz and I have discussed a
solution which he will implement).  Specifically, code from
driver.jelly (which is not yet part of a goal) is responsible for
setting the "sourcesPresent" property.  For Fulcrum, which uses a
"pom.build.sourceDirectory" directory which is created in its target/
dir, the directory is not present at the time driver.jelly is
executed, meaning that the "sourcesPresent" is not set, meaning that
Fulcrum's sources are not compiled the first time the build is run
after a "clean".  You can work around this manually by doing a `mkdir
-p target/src/`, or by executing Fulcrum's Maven build twice.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222585 13f79535-47bb-0310-9956-ffa450edef68