initial conversion to avalon components.
the components will still use the old configuration file which uses properties
as well as the standard xml configuration.

modified TurbineServices so that if it is used along with FulcrumContainer,
an application will have very little code to modify for this first cut.

the only service functionality that was altered was the cache service which
was modified so the housekeeping thread will be stopped during disposal.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222589 13f79535-47bb-0310-9956-ffa450edef68